train/requirements.txt
cog==0.8.6
decord==0.6.0
fairscale==0.4.4
hnswlib==0.7.0
lmdb==1.4.1
matplotlib
numpy
opencv_python
opencv_python_headless
pandas
Pillow
pycocoevalcap
pycocotools
PyYAML==6.0.1
ruamel_yaml_conda==0.15.100
scipy==1.9.0
soyclustering==0.2.0
timm==0.4.12
torch==1.12.1+cu116
torchvision==0.13.1+cu116
transformers==4.25.1
